+++ agent/listener/psapi.py
@@ -1,9 +1,11 @@
 import psutil as ps
 import os
 import logging
+import datetime
 import re
 import platform
-from nodes import ParentNode, RunnableNode, LazyNode
+import server
+from nodes import ParentNode, RunnableNode, RunnableParentNode, LazyNode
 from pluginnodes import PluginAgentNode
 import services
 import processes
@@ -14,6 +16,11 @@ importables = (
     'windowscounters',
     'windowslogs'
 )
+
+def get_uptime():
+    current_time = datetime.datetime.now().strftime('%s')
+    epoch_boot = int(current_time)
+    return ([epoch_boot - ps.BOOT_TIME], 's')
 
 
 def make_disk_nodes(disk_name):
@@ -39,7 +46,10 @@ def make_mountpoint_nodes(partition_name):
     used_percent = RunnableNode('used_percent', method=lambda: (ps.disk_usage(mountpoint).percent, '%'))
     device_name = RunnableNode('device_name', method=lambda: ([partition_name.device], 'name'))
     safe_mountpoint = re.sub(r'[\\/]+', '|', mountpoint)
-    return ParentNode(safe_mountpoint, children=[total_size, used, free, used_percent, device_name])
+    return RunnableParentNode(safe_mountpoint,
+                              children=[total_size, used, free, used_percent, device_name],
+                              primary='used_percent',
+                              include=('total_size', 'used', 'free', 'used_percent'))
 
 
 def make_if_nodes(if_name):
@@ -64,7 +74,9 @@ def get_system_node():
     sys_version = RunnableNode('version', method=lambda: ([platform.uname()[3]], 'name'))
     sys_machine = RunnableNode('machine', method=lambda: ([platform.uname()[4]], 'name'))
     sys_processor = RunnableNode('processor', method=lambda: ([platform.uname()[5]], 'name'))
-    return ParentNode('system', children=[sys_system, sys_node, sys_release, sys_version, sys_machine, sys_processor])
+    sys_uptime = RunnableNode('uptime', method=get_uptime)
+    sys_agent = RunnableNode('agent_version', method=lambda: ([server.__VERSION__], ''))
+    return ParentNode('system', children=[sys_system, sys_node, sys_release, sys_version, sys_machine, sys_processor, sys_uptime, sys_agent])
 
 
 def get_cpu_node():
@@ -80,15 +92,16 @@ def get_memory_node():
     mem_virt_total = RunnableNode('total', method=lambda: (ps.virtual_memory().total, 'b'))
     mem_virt_available = RunnableNode('available', method=lambda: (ps.virtual_memory().available, 'b'))
     mem_virt_percent = RunnableNode('percent', method=lambda: (ps.virtual_memory().percent, '%'))
-    mem_virt_used = RunnableNode('used', method=lambda: (ps.virtual_memory().used, 'b'))
-    mem_virt_free = RunnableNode('free', method=lambda: (ps.virtual_memory().free, 'b'))
-    mem_virt = ParentNode('virtual',
+    mem_virt_used = RunnableNode('used', method=lambda: (ps.virtual_memory().used, 'B'))
+    mem_virt_free = RunnableNode('free', method=lambda: (ps.virtual_memory().free, 'B'))
+    mem_virt = RunnableParentNode('virtual',
+                          primary='percent',
                           children=(mem_virt_total, mem_virt_available, mem_virt_free, mem_virt_percent, mem_virt_used))
     mem_swap_total = RunnableNode('total', method=lambda: (ps.swap_memory().total, 'b'))
     mem_swap_percent = RunnableNode('percent', method=lambda: (ps.swap_memory().percent, '%'))
-    mem_swap_used = RunnableNode('used', method=lambda: (ps.swap_memory().used, 'b'))
-    mem_swap_free = RunnableNode('free', method=lambda: (ps.swap_memory().free, 'b'))
-    mem_swap = ParentNode('swap', children=[mem_swap_total, mem_swap_free, mem_swap_percent, mem_swap_used])
+    mem_swap_used = RunnableNode('used', method=lambda: (ps.swap_memory().used, 'B'))
+    mem_swap_free = RunnableNode('free', method=lambda: (ps.swap_memory().free, 'B'))
+    mem_swap = RunnableParentNode('swap', primary='percent', children=[mem_swap_total, mem_swap_free, mem_swap_percent, mem_swap_used])
     return ParentNode('memory', children=[mem_virt, mem_swap])