+++ agent/listener/services.py
@@ -1,9 +1,10 @@
 import nodes
 import platform
+import re
 import subprocess
 import tempfile
 import os
-
+from stat import ST_MODE,S_IXUSR,S_IXGRP,S_IXOTH
 
 def filter_services(m):
     def wrapper(*args, **kwargs):
@@ -33,15 +34,25 @@ class ServiceNode(nodes.LazyNode):
 
     def get_service_method(self, *args, **kwargs):
         uname = platform.uname()[0]
+        # look for systemd
+        is_systemctl = False
         try:
             process = subprocess.Popen(['which', 'systemctl'], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
             process.wait()
             if process.returncode == 0:
                 is_systemctl = True
-            else:
-                is_systemctl = False
         except:
-            is_systemctl = False
+            pass
+
+        # look for upstart
+        is_upstart = False
+        try:
+            process = subprocess.Popen(['which', 'initctl'], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            process.wait()
+            if process.returncode == 0:
+                is_upstart = True
+        except:
+            pass
 
         if uname == 'Windows':
             return self.get_services_via_sc
@@ -50,7 +61,10 @@ class ServiceNode(nodes.LazyNode):
         else:
             if is_systemctl:
                 return self.get_services_via_systemctl
+            elif is_upstart:
+                return self.get_services_via_initctl
             else:
+                # fall back on sysv init
                 return self.get_services_via_initd
 
     @filter_services
@@ -93,11 +107,48 @@ class ServiceNode(nodes.LazyNode):
 
     @filter_services
     def get_services_via_systemctl(self, *args, **kwargs):
-        return {}
+        services = {}
+        status = tempfile.TemporaryFile()
+        service = subprocess.Popen(['systemctl', 'list-units', '*.service', '--no-pager', '--no-legend'], stdout=status)
+        service.wait()
+        status.seek(0)
+
+        for line in status.readlines():
+            unit, load, active, sub, description = line.split(None, 5)
+            if active.lower() == 'active' and sub.lower() == 'running':
+                services[unit] = 'running'
+            else:
+                services[unit] = 'stopped'
+        return services
+
+    @filter_services
+    def get_services_via_initctl(self, *args, **kwargs):
+        services = {}
+        # ubuntu supports both sysv init and upstart, let upstart win
+        services = self.get_services_via_initd(args, kwargs)
+
+        # now go ask initctl
+        status = tempfile.TemporaryFile()
+        service = subprocess.Popen(['initctl', 'list'], stdout=status)
+        service.wait()
+        status.seek(0)
+
+        for line in status.readlines():
+            m = re.match("(.*) (?:\w*)/(\w*)(?:, .*)?", line)
+            try:
+                print m.groups()
+                if m.group(2) == 'running':
+                    services[m.group(1)] = 'running'
+                else:
+                    services[m.group(1)] = 'stopped'
+            except:
+                pass
+        return services
 
     @filter_services
     def get_services_via_initd(self, *args, **kwargs):
-        possible_services = os.listdir('/etc/init.d')
+        # only look at executable files in init.d (there is no README service)
+        possible_services = filter(lambda x: os.stat('/etc/init.d/'+x)[ST_MODE] & (S_IXUSR|S_IXGRP|S_IXOTH), os.listdir('/etc/init.d'))
         services = {x: 'stopped' for x in possible_services}
         devnull = open(os.devnull, 'w')