+++ agent/passive/ncpacheck.py
@@ -7,17 +7,20 @@ import listener.server
 
 class NCPACheck(object):
     """
-    This class is meant to simplify the creation of other classes that simply run a
-    command and return it to Nagios is some format. The immediate motivation for this
-    is the difference between NSCA and NRDP. They share a lot of functionality that would
-    be otherwise duplicated and should not be in the abstract class.
+    This class is meant to simplify the creation of other classes that
+    simply run a command and return it to Nagios is some format. The
+    immediate motivation for this is the difference between NSCA and
+    NRDP. They share a lot of functionality that would be otherwise
+    duplicated and should not be in the abstract class.
 
-    Convenience class for wrapping up the commands defined in the config and setting their host/service to be accessed
-    by their child classes and running the results.
+    Convenience class for wrapping up the commands defined in the
+    config and setting their host/service to be accessed by their
+    child classes and running the results.
     """
 
-    def __init__(self, instruction, hostname, servicename):
+    def __init__(self, config, instruction, hostname, servicename):
         logging.debug('Initializing NCPA check with %s', instruction)
+        self.config = config
         self.hostname = hostname
         self.servicename = servicename
         self.instruction = instruction
@@ -29,7 +32,8 @@ class NCPACheck(object):
 
         :param instruction: The instruction to be parsed
         :type instruction: unicode
-        :return: Tuple containing the API URL that will be accessed to retrieve information from the local agent and its args
+        :return: Tuple containing the API URL that will be accessed to retrieve
+                 information from the local agent and its args
         :rtype:  tuple
         """
         logging.debug('Getting API url for instruction %s', instruction)
@@ -48,12 +52,13 @@ class NCPACheck(object):
 
     def run(self):
         """
-        The primary method for running an NCPA check. Once the method has been instantiated
-        this is the only function you should run.
+        The primary method for running an NCPA check. Once the method has been
+        instantiated this is the only function you should run.
 
         :return: A tuple containing the stdout and returncode of the NCPA check
         :rtype: tuple
-        :raise: Will raise ValueError when either the stdout or returncode are not a string
+        :raise: Will raise ValueError when either the stdout or returncode are
+                not a string
         :raises: ValueError
         """
         logging.info("Running check: %s", self.instruction)
@@ -63,12 +68,12 @@ class NCPACheck(object):
         stdout, returncode = self.handle_agent_response(response)
 
         if stdout is None or returncode is None:
-            raise ValueError("Stdout or returncode was None, cannot return meaningfully.")
+            raise ValueError("Stdout or returncode was None, cannot return "
+                             "meaningfully.")
 
         return stdout, returncode
 
-    @staticmethod
-    def run_check(api_url, api_args):
+    def run_check(self, api_url, api_args):
         """
         Access the local agent using api_url
 
@@ -76,8 +81,9 @@ class NCPACheck(object):
         :type api_args: dict
         :param api_url: API URL that will be asked for the NCPA check
         :type api_url: unicode
-        :return: A dict containing the standard out and return code of the check
-        :rtype: dict
+        :return: A dict containing the standard out and return code of the
+                 check
+        :rtype: str
         """
         query = urllib.urlencode(api_args)
         complete_api_url = "{}?{}".format(api_url, query)
@@ -85,6 +91,7 @@ class NCPACheck(object):
         logging.debug("Access the API with %s", complete_api_url)
 
         listener.server.__INTERNAL__ = True
+        listener.server.listener.config['iconfig'] = self.config
         api_server = listener.server.listener.test_client()
 
         try:
@@ -113,13 +120,16 @@ class NCPACheck(object):
             stdout = response_dict['stdout']
             returncode = unicode(response_dict['returncode'])
         except ValueError as exc:
-            logging.error("Error with JSON: %s. JSON was: %s", str(exc), response)
+            logging.error("Error with JSON: %s. JSON was: %s", str(exc),
+                          response)
         except TypeError as exc:
             logging.error("Error response was not a string: %s", str(exc))
         except KeyError as exc:
-            logging.error("JSON was missing keyword: %s. JSON given: %s", str(exc), response)
+            logging.error("JSON was missing keyword: %s. JSON given: %s",
+                          str(exc), response)
 
-        logging.debug("JSON response handled found stdout='%s', returncode=%s", stdout, returncode)
+        logging.debug("JSON response handled found stdout='%s', returncode=%s",
+                      stdout, returncode)
         return stdout, returncode
 
     @staticmethod
@@ -131,12 +141,15 @@ class NCPACheck(object):
 
         As opposed to the URL encoded version
 
-        :param instruction: The config entry that needs to parsed as a command instruction
+        :param instruction: The config entry that needs to parsed as a command
+                            instruction
         :type instruction: unicode
-        :return: tuple containing the API URL and a dictionary containing the arguments
+        :return: tuple containing the API URL and a dictionary containing the
+                 arguments
         :rtype: tuple
         """
-        logging.debug("Parsing command line style instruction: %s", instruction)
+        logging.debug("Parsing command line style instruction: %s",
+                      instruction)
         stripped_instruction = instruction.strip().split(' ')
 
         api_url = stripped_instruction[0]
@@ -144,8 +157,9 @@ class NCPACheck(object):
 
         arguments = stripped_instruction[1:]
 
-        # Parse each individual variable, this will go through each command line argument and get the argument names
-        # and their individual variables.
+        # Parse each individual variable, this will go through each command
+        # line argument and get the argument names and their individual
+        # variables.
         while arguments:
 
             argument = arguments.pop(0)
@@ -156,11 +170,14 @@ class NCPACheck(object):
                     arg_name = argument
                     arg_value = arguments.pop(0)
                 except IndexError:
-                    logging.warning("Unable to parse all arguments from instruction. Mis-paired option: %s", argument)
+                    logging.warning("Unable to parse all arguments from "
+                                    "instruction. Mis-paired option: %s",
+                                    argument)
                     break
 
-            # We need to strip off the leading '--' or '-' if they exist
-            # as we are using these flags as the literal variables we want to know about.
+            # We need to strip off the leading '--' or '-' if they exist as we
+            # are using these flags as the literal variables we want to know
+            # about.
             if arg_name.startswith('--'):
                 arg_name = arg_name[2:]
             elif arg_name.startswith('-'):
@@ -173,8 +190,9 @@ class NCPACheck(object):
     @staticmethod
     def normalize_api_url(api_url):
         """
-        Due to legacy concerns, we must support both /api/cpu/percent and cpu/percent. Notice the api_url passed to
-        the actual listener must start with /api.
+        Due to legacy concerns, we must support both /api/cpu/percent and
+        cpu/percent. Notice the api_url passed to the actual listener must
+        start with /api.
 
         :param api_url: The un-normalized api URL.
         :type api_url: unicode
@@ -196,9 +214,9 @@ class NCPACheck(object):
 
     @staticmethod
     def parse_api_url_style_instruction(instruction):
-        parsed = urlparse.urlparse(instruction)
+        parse = urlparse.urlparse(instruction)
 
-        api_url = parsed.path
-        api_args = {x: v[0] for x, v in urlparse.parse_qs(parsed.query).items()}
+        api_url = parse.path
+        api_args = {x: v[0] for x, v in urlparse.parse_qs(parse.query).items()}
 
         return api_url, api_args